API Reference

API Reference

ElasticFDSG.dim2.configtemplate

ElasticFDSG.dim2.runsim

ElasticFDSG.dim3.configtemplate

ElasticFDSG.dim3.runsim

ElasticFDSG.dim2.configtemplate — Method

Elastic_FDSG.dim2.configtemplate(path)

Creates and empty template configuration file for a 2D ElasticFDSG simulation. The user must fill the template afterwards.

Arguments

path::String: Path where the template is saved.

Returns

Nothing: The function saves the template as specified in path.

ElasticFDSG.dim2.runsim — Method

Elastic_FDSG.dim2.runsim(CONFIGPATH, VELMODPATH)

Run the 2D elastic forward simulation using the specified configuration and velocity model.

Arguments

CONFIGPATH::String: Path to the configuration.yaml file containing simulation settings.

VELMODPATH::String: Path to the velocity model file. Supported formats include .jld2, .npy, and .npz.

Returns

Nothing: The function runs the simulation and saves the results as specified in the configuration file.

ElasticFDSG.dim3.configtemplate — Method

Elastic_FDSG.dim3.configtemplate(path)

Creates and empty template configuration file for a 3D ElasticFDSG simulation. The user must fill the template afterwards.

Arguments

path::String: Path where the template is saved.

Returns

Nothing: The function saves the template as specified in path.

ElasticFDSG.dim3.runsim — Method

Elastic_FDSG.dim3.runsim(CONFIGPATH, VELMODPATH)

Run the 3D elastic forward simulation using the specified configuration and velocity model.

Arguments

CONFIGPATH::String: Path to the configuration.yaml file containing simulation settings.

VELMODPATH::String: Path to the velocity model file. Supported formats include .jld2, .npy, and .npz.

Returns

Nothing: The function runs the simulation and saves the results as specified in the configuration file.