API Reference

API Reference

ElasticFDSG.config_template

ElasticFDSG.dim2.runsim

ElasticFDSG.dim3.runsim

ElasticFDSG.load_results

ElasticFDSG.print_h5_tree

ElasticFDSG.load_results — Function

ElasticFDSG.load_results(filename) Loads a result.h5 file into a dictionary.

Arguments

filename::Stringg: Path to the results.h5 file.

ElasticFDSG.print_h5_tree — Function

ElasticFDSG.printh5tree(filename) Prints tree structure of a result.h5 file.

Arguments

filename::Stringg: Path to the results.h5 file.

ElasticFDSG.config_template — Function

ElasticFDSG.config_template(path, dim)

Creates and saves an empty template configuration.yaml file for a ElasticFDSG simulation. Users can fill the template afterwards.

Arguments

path::String: Path where the template is saved.

dim::Real: dim = 2 => 2D-template, dim=3 => 3D-template

ElasticFDSG.dim2.runsim — Function

ElasticFDSG.dim2.runsim(CONFIGPATH, VELMODPATH) Run the 2D elastic forward simulation using the specified configuration and velocity model.

Arguments:

CONFIGPATH::String: Path to the configuration.yaml file containing simulation settings.

VELMODPATH::String: Path to the velocity model file. Supported formats: .jld2, .npy, and .npz.

Optional Keyword Arguments:

block_size::Tuple=(32,32) : Specify numbers of threads launched on GPU. For CPU it will have no effect.

return_::Bool=false: False -> Results will be saved as .h5 file as specified in the config file. True -> Function returns FDSG2D struct.

ElasticFDSG.dim3.runsim — Function

ElasticFDSG.dim3.runsim(CONFIGPATH, VELMODPATH) Run the 3D elastic forward simulation using the specified configuration and velocity model.

Arguments:

CONFIGPATH::String: Path to the configuration.yaml file containing simulation settings.

VELMODPATH::String: Path to the velocity model file. Supported formats: .jld2, .npy, and .npz.

Optional Keyword Arguments:

block_size::Tuple=(7,7,7) : Specify numbers of threads launched on GPU. For CPU it will have no effect.

return_::Bool=false: False -> Results will be saved as .h5 file as specified in the config file. True -> Function returns FDSG3D struct.